Router Pattern/requirements.txt
google-adk>=0.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
chromadb>=0.4.0

# For real Teams/Email notifications
requests>=2.31.0
httpx>=0.25.0
